+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 **Release Date**: 28th September 2022
 ### Features
 1. Add import button for Instance SSH Key.
-2. Add Network Firewall Resource.
+2. Add Network Firewall Resource (Currently no Terraform Generation).
 3. Add Variables View allowing specification of Terraform Variables that can be used within the design.
 4. Update Cost Estimator to include Network Firewall.
 ### Bug Fixes